add blocks
need to figure out scoping. the languages that I know of that have implicit
declaration have a single global scope, where sub-blocks can access and modify
within it. lua includes functions in sub-blocks, able to freely access and
modify the global scope; python shadows variables if they are assigned to in a
function, but reassigns it if it's assigned to in other types of blocks.

currently I'm thinking functions will have isolated envs, while a normal block
will re-assign to a parent if the parent contains it or declare if not.