turns out I don't actually want java-style pointer fucking
when I push the closure, it also pushes the interp's env, and vice-versa when I
pop the closure, so when make_counter is interpreted, I get a refcounted
pointer to the interp's env as a closure, where it eventually pops the actual
closure that I'd like to keep around.

it's nice when you want that behavior, but I don't think I do here. I'll have
to do some more research and see if the java implementation deep-clones when
creating a function's closure, and whether or not that's feasable with
rc/refcell.