fuck
turns out that I wasn't actually done. I still need to add the important part
of methods, i.e. the implicit parameter, but I think I'm fucked. this is the
exact line when I realized this:

    We need to take this at the point that the method is accessed and
    attach it to the function somehow so that it stays around as long as
    we need it to. Hmm… a way to store some extra data that hangs around
    a function, eh? That sounds an awful lot like a closure, doesn't it?

I'm at least 70% sure that I'm totally screwed. I'll need to add it as
syntactic sugar, passing the object that we got it from to the function. I'm
thinking maybe in visit_get I can hold on to whatever object evals to, and
pass it to the call. if that doesn't work...

looks like I'm back to writing esolangs.