filter newlines
I *think* the syntax is unambiguous enough to allow for removing newlines?
I'll need to do some more testing to find out. maybe split out consume to
one that ignores newlines and one that doesn't. there's still some funkiness
around `retn` that might be helped by this, but I haven't touched that in a
while.