cargo fmt, clean
this might actually make it easier to scan on demand and not have to
deal with lifetime issues. I didn't realize earlier that
identifier_constant always is passed the previous token instead of
whatever token the caller wants. this probably isn't true when we do
local variable resolution but when we do that it'll be easy enough to
just have that in-line.