make scanner's tokens a Vec<Option<Token>>
this is so we can use Option::take to maybe satisfy the borrow checker.
the current Scanner::current and Scanner::previous return a reference so
it's a little unwieldy when we want to steal tokens from the scanner, and
by nature it's *very* non-lexical which the compiler can't wrap its head
around just yet.

it's a bit of a gamble but I think it'll work out...