update justfile, readme
also do some weird pointer stuff?? I'm not sure how cross-platform this is or
even if it's safe (which it's probably not). it at least seems to be type safe
since there's no way to assign to a trait object directly, and downcast has
some safeguards to prevent crazy stuff from happening. you could probably still
do transmute on it though...