make a note to remove visitor pattern stuff
tbh I think the mutually recursive function pattern that passes around a &mut
Scanner in the parser module was probably a better idea than having an object
that holds on to a &mut Scanner. I think refactoring the emitter in the same
way is going to give me the same benefits.