I'm genuinely surprised I found this and didn't go crazy
so when I ran with --features "pc-debug" i was getting errors related to local
variables. this made me immediately sad because I'm in the process of changing a
whole lot of stuff with the vm to make it reentrant. my first thought was a call
to debug! somewhere was mutating some state that I had been messing with, which
would definitely have driven me off the deep end. then the holy spirit itself
guided my fingers to type rg 'pc-debug' and i found this???? how did this
happen???????? anyway it's fixed and I'm still sane, and that's the improtant
part